SCons: Do not enforce clang on posix, allowing use of GCC
The compiler can now be defined with `scons CC=clang CXX=clang++`
and/or with the `CC` and `CXX` environment variables.

The build will pick the default C and C++ compilers of the user's
system, so probably GCC on most, and Clang and some of them (e.g. on
Travis which is configured to use Clang).

Also cleaned up unnecessary definitions which are already part of the
base environment, and removed the unnecessary (default on Linux) and
Clang-only -stdlib=libstdc++ definition (still needed on macOS).