Recharge: Melt chance and shop customization
For the recharge spell, there is now a chance for the item to melt if you try to charge it too much.
For recharge shops, it is now possible to configure how much they are willing to charge an item,
as well as how many charges are given per payment.