Remove warning for spells that use default 1-sq pattern.
This fixes #489.

I considered an alternate solution of adding a case for every 1-sq-target spell, but there are so many of them. Targeting 1 square is a sensible default, not something that should have to be specified, and this warning would only really be useful when adding new spells to avoid forgetting to specify their shape. I say we don't need it.